+++ confidence/io.py
@@ -1,28 +1,28 @@
 import logging
 import re
 import typing
+import warnings
 from enum import IntEnum
 from functools import partial
 from itertools import product
 from os import PathLike, environ, pathsep
 from pathlib import Path
 
-import yaml
-
-from confidence.models import Configuration, Missing, NoDefault, NotConfigured, unwrap
+from confidence.formats import YAML, Format
+from confidence.models import Configuration, Missing, NoDefault, NotConfigured
 
 
 LOG = logging.getLogger(__name__)
 
 
-def read_xdg_config_dirs(name: str, extension: str) -> Configuration:
+def read_xdg_config_dirs(name: str, format: Format = YAML) -> Configuration:
     """
     Read from files found in XDG-specified system-wide configuration paths,
     defaulting to ``/etc/xdg``. Depends on ``XDG_CONFIG_DIRS`` environment
     variable.
 
     :param name: application or configuration set name
-    :param extension: file extension to look for
+    :param format: configuration (file) format to use
     :returns: a `Configuration` instance with values read from XDG-specified
         directories
     """
@@ -32,29 +32,34 @@ def read_xdg_config_dirs(name: str, extension: str) -> Configuration:
     config_dirs = reversed(config_dirs.split(pathsep))
 
     # load a file from all config dirs, default to NotConfigured
-    return loadf(*(Path(config_dir) / f'{name}.{extension}' for config_dir in config_dirs), default=NotConfigured)
+    return loadf(
+        *(Path(config_dir) / f'{name}{format.suffix}' for config_dir in config_dirs),
+        format=format,
+        default=NotConfigured,
+    )
 
 
-def read_xdg_config_home(name: str, extension: str) -> Configuration:
+def read_xdg_config_home(name: str, format: Format = YAML) -> Configuration:
     """
     Read from file found in XDG-specified configuration home directory,
     expanding to ``${HOME}/.config/name.extension`` by default. Depends on
     ``XDG_CONFIG_HOME`` or ``HOME`` environment variables.
 
     :param name: application or configuration set name
-    :param extension: file extension to look for
+    :param format: configuration (file) format to use
     :returns: a `Configuration` instance, possibly `NotConfigured`
     """
     # find optional value of ${XDG_CONFIG_HOME}
     # XDG spec: "If $XDG_CONFIG_HOME is either not set or empty, a default equal to $HOME/.config should be used."
-    # see https://specifications.freedesktop.org/basedir-spec/latest/ar01s03.html
+    # see https://specifications.freedesktop.org/basedir-spec/latest/
+    home = environ.get('HOME')
     config_home = environ.get('XDG_CONFIG_HOME')
-    config_home = Path(config_home) if config_home else Path('~/.config').expanduser()
+    config_home = Path(config_home) if config_home else Path(f'{home}/.config')
     # expand to full path to configuration file in XDG config path
-    return loadf(config_home / f'{name}.{extension}', default=NotConfigured)
+    return loadf(config_home / f'{name}{format.suffix}', format=format, default=NotConfigured)
 
 
-def read_envvars(name: str, extension: typing.Optional[str] = None) -> Configuration:
+def read_envvars(name: str, format: Format = YAML) -> Configuration:
     """
     Read environment variables starting with ``NAME_``, where subsequent
     underscores are interpreted as namespaces. Underscores can be retained as
@@ -70,7 +75,7 @@ def read_envvars(name: str, extension: typing.Optional[str] = None) -> Configura
         `.read_envvar_file`.
 
     :param name: environment variable prefix to look for (without the ``_``)
-    :param extension: *(unused)*
+    :param format: configuration (file) format to use
     :returns: a `Configuration` instance, possibly `NotConfigured`
     """
     prefix = f'{name}_'
@@ -95,39 +100,39 @@ def read_envvars(name: str, extension: typing.Optional[str] = None) -> Configura
     LOG.info(f'reading configuration from {len(values)} {prefix}* environment variables')
 
     # pass value to yaml.safe_load to align data type transformation with reading values from files
-    return Configuration({dotted(name): yaml.safe_load(value) for name, value in values.items()})
+    return Configuration({dotted(name): format.loads(value) for name, value in values.items()})
 
 
-def read_envvar_file(name: str, extension: typing.Optional[str] = None) -> Configuration:
+def read_envvar_file(name: str, format: Format = YAML) -> Configuration:
     """
     Read values from a file provided as a environment variable
     ``NAME_CONFIG_FILE``.
 
     :param name: environment variable prefix to look for (without the
         ``_CONFIG_FILE``)
-    :param extension: *(unused)*
+    :param format: configuration (file) format to use
     :returns: a `Configuration`, possibly `NotConfigured`
     """
     envvar_file = environ.get(f'{name}_config_file'.upper())
     if envvar_file:
         # envvar set, load value as file
-        return loadf(envvar_file)
+        return loadf(envvar_file, format=format)
     else:
         # envvar not set, return an empty source
         return NotConfigured
 
 
-def read_envvar_dir(envvar: str, name: str, extension: str) -> Configuration:
+def read_envvar_dir(envvar: str, name: str, format: Format = YAML) -> Configuration:
     """
     Read values from a file located in a directory specified by a particular
-    environment file. ``read_envvar_dir('HOME', 'example', 'yaml')`` would
+    environment file. ``read_envvar_dir('HOME', 'example', format=YAML)`` would
     look for a file at ``/home/user/example.yaml``. When the environment
     variable isn't set or the file does not exist, `NotConfigured` will be
     returned.
 
     :param envvar: the environment variable to interpret as a directory
     :param name: application or configuration set name
-    :param extension: file extension to look for
+    :param format: configuration (file) format to use
     :returns: a `Configuration`, possibly `NotConfigured`
     """
     config_dir = environ.get(envvar)
@@ -135,8 +140,8 @@ def read_envvar_dir(envvar: str, name: str, extension: str) -> Configuration:
         return NotConfigured
 
     # envvar is set, construct full file path, expanding user to allow the envvar containing a value like ~/config
-    config_path = Path(config_dir).expanduser() / f'{name}.{extension}'
-    return loadf(config_path, default=NotConfigured)
+    config_path = Path(config_dir).expanduser() / f'{name}{format.suffix}'
+    return loadf(config_path, format=format, default=NotConfigured)
 
 
 class Locality(IntEnum):
@@ -152,30 +157,30 @@ class Locality(IntEnum):
     ENVIRONMENT = 3  #: configuration from environment variables
 
 
-Loadable = typing.Union[str, typing.Callable[[str, str], Configuration]]
+Loadable = typing.Union[str, typing.Callable[[str, Format], Configuration]]
 
 
 _LOADERS: typing.Mapping[Locality, typing.Iterable[Loadable]] = {
     Locality.SYSTEM: (
         # system-wide locations
         read_xdg_config_dirs,
-        '/etc/{name}/{name}.{extension}',
-        '/etc/{name}.{extension}',
-        '/Library/Preferences/{name}/{name}.{extension}',
-        '/Library/Preferences/{name}.{extension}',
+        '/etc/{name}/{name}{suffix}',
+        '/etc/{name}{suffix}',
+        '/Library/Preferences/{name}/{name}{suffix}',
+        '/Library/Preferences/{name}{suffix}',
         partial(read_envvar_dir, 'PROGRAMDATA'),
     ),
     Locality.USER: (
         # user-local locations
         read_xdg_config_home,
-        '~/Library/Preferences/{name}.{extension}',
+        '~/Library/Preferences/{name}{suffix}',
         partial(read_envvar_dir, 'APPDATA'),
         partial(read_envvar_dir, 'LOCALAPPDATA'),
-        '~/.{name}.{extension}',
+        '~/.{name}{suffix}',
     ),
     Locality.APPLICATION: (
         # application-local locations
-        './{name}.{extension}',
+        './{name}{suffix}',
     ),
     Locality.ENVIRONMENT: (
         # application-specific environment variables
@@ -199,7 +204,7 @@ def loaders(*specifiers: typing.Union[Locality, Loadable]) -> typing.Iterable[Lo
         # an explicit path, a template and a user-defined function
         load_order = loaders(Locality.user,
                              '/etc/defaults/hard-coded.yaml',
-                             '/path/to/{name}.{extension}',
+                             '/path/to/{name}{suffix}',
                              my_loader)
 
         # load configuration for name 'my-application' using the load order
@@ -228,27 +233,30 @@ DEFAULT_LOAD_ORDER = tuple(
 )
 
 
-def load(*fps: typing.IO, missing: typing.Any = Missing.SILENT) -> Configuration:
+def load(*fps: typing.TextIO, format: Format = YAML, missing: typing.Any = Missing.SILENT) -> Configuration:
     """
     Read a `Configuration` instance from file-like objects.
 
     :param fps: file-like objects (supporting ``.read()``)
+    :param format: configuration (file) format to use
     :param missing: policy to be used when a configured key is missing, either
         as a `Missing` instance or a default value
     :returns: a `Configuration` instance providing values from *fps*
     """
-    return Configuration(*(yaml.safe_load(fp.read()) for fp in fps), missing=missing)
+    return Configuration(*(format.load(fp) for fp in fps), missing=missing)
 
 
 def loadf(
     *fnames: typing.Union[str, PathLike],
+    format: Format = YAML,
     default: typing.Any = NoDefault,
     missing: typing.Any = Missing.SILENT,
 ) -> Configuration:
     """
     Read a `Configuration` instance from named files.
 
-    :param fnames: name of the files to ``open()``
+    :param fnames: name of the files to read
+    :param format: configuration (file) format to use
     :param default: `dict` or `Configuration` to use when a file does not
         exist (default is to raise a `FileNotFoundError`)
     :param missing: policy to be used when a configured key is missing, either
@@ -258,10 +266,7 @@ def loadf(
 
     def readf(fpath: Path) -> typing.Mapping[str, typing.Any]:
         try:
-            with fpath.open('r') as fp:
-                LOG.info(f'reading configuration from file {fpath}')
-                # default to empty dict, yaml.safe_load will return None for an empty document
-                return yaml.safe_load(fp.read()) or {}
+            return format.loadf(fpath)
         except OSError:
             # file does not exist or inaccessible
             if default is NoDefault:
@@ -271,26 +276,29 @@ def loadf(
                 LOG.debug(f'unable to read configuration from file {fpath}')
                 return default
 
+    # expand the user directories here, format is not in charge of the file paths
     return Configuration(*(readf(Path(fname).expanduser()) for fname in fnames), missing=missing)
 
 
-def loads(*strings: str, missing: typing.Any = Missing.SILENT) -> Configuration:
+def loads(*strings: str, format: Format = YAML, missing: typing.Any = Missing.SILENT) -> Configuration:
     """
     Read a `Configuration` instance from strings.
 
     :param strings: configuration contents
+    :param format: configuration (file) format to use
     :param missing: policy to be used when a configured key is missing, either
         as a `Missing` instance or a default value
     :returns: a `Configuration` instance providing values from *strings*
     """
-    return Configuration(*(yaml.safe_load(string) for string in strings), missing=missing)
+    return Configuration(*(format.loads(string) for string in strings), missing=missing)
 
 
 def load_name(
     *names: str,
     load_order: typing.Iterable[Loadable] = DEFAULT_LOAD_ORDER,
-    extension: str = 'yaml',
+    format: Format = YAML,
     missing: typing.Any = Missing.SILENT,
+    extension: None = None,  # NB: parameter is deprecated, see below
 ) -> Configuration:
     """
     Read a `Configuration` instance by name, trying to read from files in
@@ -305,61 +313,80 @@ def load_name(
         significance
     :param load_order: ordered list of name templates or `callable` s, in
         increasing order of significance
-    :param extension: file extension to be used
+    :param format: configuration (file) format to use
     :param missing: policy to be used when a configured key is missing, either
         as a `Missing` instance or a default value
     :returns: a `Configuration` instances providing values loaded from *names*
         in *load_order* ordering
     """
+    if extension is not None:
+        if format is YAML:
+            warnings.warn(
+                'extension argument to load_name has been deprecated, use the format argument to set the file suffix',
+                category=DeprecationWarning,
+                stacklevel=2,  # warn about user code calling load_name rather than load_name itself`
+            )
+            format = YAML(suffix=f'.{extension}')
+        else:
+            raise ValueError("format and extension cannot be combined, use format's suffix")
 
     def generate_sources() -> typing.Iterable[typing.Mapping[str, typing.Any]]:
         # argument order for product matters, for names "foo" and "bar":
         # /etc/foo.yaml before /etc/bar.yaml, but both of them before ~/.foo.yaml and ~/.bar.yaml
         for source, name in product(load_order, names):
             if callable(source):
-                yield source(name, extension)
+                yield source(name, format)
             else:
-                # expand user to turn ~/.name.yaml into /home/user/.name.yaml
-                candidate = Path(source.format(name=name, extension=extension)).expanduser()
-                yield loadf(candidate, default=NotConfigured)
+                candidate = Path(source.format(name=name, suffix=format.suffix))
+                yield loadf(candidate, format=format, default=NotConfigured)
 
     return Configuration(*generate_sources(), missing=missing)
 
 
-def dump(value: typing.Any, fp: typing.IO, encoding: str = 'utf-8') -> None:
+def _check_format_encoding(format: Format, encoding: typing.Optional[str]) -> Format:
+    if encoding:
+        if format is YAML:
+            warnings.warn(
+                'encoding argument to dump functions has been deprecated, use the format argument to set the encoding',
+                category=DeprecationWarning,
+                stacklevel=3,
+            )
+            return format(encoding=encoding)
+        else:
+            raise ValueError("format and encoding cannot be combined, use format's encoding")
+
+    return format
+
+
+def dump(
+    value: typing.Any,
+    fp: typing.TextIO,
+    format: Format = YAML,
+    encoding: None = None,  # NB: parameter is deprecated, see _check_format_encoding
+) -> None:
     """
-    Serialize the configuration in *value* to YAML format, writing it to *fp*.
-
-    :param value: the value (like a `Configuration` object) to dump
-    :param fp: a file-like object to write to
-    :param encoding: encoding to use
+    Shorthand for `format.dump(value, fp)`.
     """
-    # recursively unwrap the value to help yaml understand what we're trying to dump
-    # use block style output for nested collections (flow style dumps nested dicts inline)
-    yaml.safe_dump(unwrap(value), stream=fp, encoding=encoding, default_flow_style=False)
+    format = _check_format_encoding(format, encoding)
+    format.dump(value, fp)
 
 
-def dumpf(value: typing.Any, fname: typing.Union[str, PathLike], encoding: str = 'utf-8') -> None:
+def dumpf(
+    value: typing.Any,
+    fname: typing.Union[str, PathLike],
+    format: Format = YAML,
+    encoding: None = None,  # NB: parameter is deprecated, see _check_format_encoding
+) -> None:
     """
-    Serialize the configuration in *value* to a YAML-formatted file.
-
-    :param value: the value (like a `Configuration` object) to dump
-    :param fname: name or path of the file to write to
-    :param encoding: encoding to use
+    Shorthand for `format.dumpf(value, fname)`.
     """
-    with Path(fname).open('wb') as out_file:
-        dump(value, out_file, encoding=encoding)
+    format = _check_format_encoding(format, encoding)
+    # expand the dump path here, format is not in charge of the file paths
+    format.dumpf(value, Path(fname).expanduser())
 
 
-def dumps(value: typing.Any) -> str:
+def dumps(value: typing.Any, format: Format = YAML) -> str:
     """
-    Serialize the configuration in *value* as a YAML-formatted string.
-
-    :param value: the value (like a `Configuration` object) to dump
-    :returns: *configuration*, serialized as a `str` in YAML format
+    Shorthand for `format.dumps(value)`.
     """
-    # recursively unwrap the value to help yaml understand what we're trying to dump
-    # use block style output for nested collections (flow style dumps nested dicts inline)
-    encoded = yaml.safe_dump(unwrap(value), default_flow_style=False)
-    # omit explicit document end (...) included with simple values
-    return encoded.removesuffix('\n...\n')
+    return format.dumps(value)